+++ foxhole_stockpiles/commands/update_db/__init__.py
@@ -1,5 +1,0 @@
-"""Database update command for migrating template databases."""
-
-from foxhole_stockpiles.commands.update_db.update_db import main
-
-__all__ = ["main"]